+++ changelog.md
@@ -2,15 +2,20 @@
 Notable changes to the `cc.isr.vi` solution are documented in this file using the 
 [Keep a Changelog] style. The dates specified are in coordinated universal time (UTC).
 
-[8.1.9455]: https://www.github.com/atecoder/dn.vi.ivi
+[8.1.9456]: https://www.github.com/atecoder/dn.vi.ivi
 
 Current IVI Visa Compatibility: 8.0.1.9391
+
+## [8.1.9456] - 2025-11-21
+- Add Path extensions get temp folder.
+- Add ignore ending to file comparizones and rename methods.
 
 ## [8.1.9455] - 2025-11-20
 - Access subsystem base: update implementation.
 Exception extensions:
 - Add option to add stack trace to the message.
 - Add message to the trace of the native exception.
+- In Try methods: report message details if IO or Invalid operation exception; otherwise build a full message.
 
 ## [8.1.9454] - 2025-11-19
 Device Tsp